+++ components/deploy/StaticPreview.tsx
@@ -1,111 +1,8 @@
 import React, { useEffect, useRef, useState } from 'react';
 import { useDeployContext } from '@/providers/DeployProvider';
-
-function parseImports(code: string) {
-  // Matches: import X from 'pkg'; import {A, B} from 'pkg'; import * as X from 'pkg';
-  // Also matches local, CSS, and type-only imports
-  // Updated regex to match scoped packages (e.g., @radix-ui/react-slot)
-  const importRegex = /import\s+((\w+)|\*\s+as\s+(\w+)|\{([^}]+)\})?\s*from\s*['"]([^'"]+)['"];?|import\s+['"]([^'"]+\.css)['"];?|import\s+type\s+\{[^}]+\}\s+from\s+['"]([^'"]+)['"];?/g;
-  let match;
-  const imports: Array<{
-    defaultImport?: string;
-    namedImports?: string[];
-    namespaceImport?: string;
-    package: string;
-    isLocal: boolean;
-    importStatement: string;
-    importedAs?: string;
-    isTypeOnly?: boolean;
-    isCss?: boolean;
-  }> = [];
-  let codeWithoutImports = code;
-  while ((match = importRegex.exec(code)) !== null) {
-    const [full, group, defaultImport, namespaceImport, namedImports, pkg, cssImport, typeImport] = match;
-    // CSS import
-    if (cssImport) {
-      codeWithoutImports = codeWithoutImports.replace(full, '');
-      continue;
-    }
-    // Type-only import
-    if (typeImport) {
-      codeWithoutImports = codeWithoutImports.replace(full, '');
-      continue;
-    }
-    // Local import (starts with ./, ../, or @/)
-    if (pkg && (pkg.startsWith('./') || pkg.startsWith('../') || pkg.startsWith('@/'))) {
-      codeWithoutImports = codeWithoutImports.replace(full, '');
-      if (namespaceImport) {
-        imports.push({ namespaceImport, package: pkg, isLocal: true, importStatement: full });
-      } else if (namedImports) {
-        imports.push({ namedImports: namedImports.split(',').map(s => s.trim().split(' as ')[0]), package: pkg, isLocal: true, importStatement: full });
-      } else if (defaultImport) {
-        imports.push({ defaultImport, package: pkg, isLocal: true, importStatement: full });
-      }
-      continue;
-    }
-    // NPM import (scoped or unscoped)
-    if (pkg) {
-      imports.push({
-        defaultImport,
-        namedImports: namedImports ? namedImports.split(',').map(s => s.trim().split(' as ')[0]) : undefined,
-        namespaceImport,
-        package: pkg,
-        isLocal: false,
-        importStatement: full
-      });
-    }
-  }
-  return { imports, codeWithoutImports };
-}
-
-function getAllNpmImports(files: Array<{ file: string; data: string }>) {
-  // Collect all npm package imports from all files
-  const npmPkgs = new Set<string>();
-  files.forEach(f => {
-    const { imports } = parseImports(f.data);
-    imports.forEach(imp => {
-      if (!imp.isLocal && imp.package && !imp.package.startsWith('.') && !imp.package.startsWith('@/')) {
-        npmPkgs.add(imp.package);
-      }
-    });
-  });
-  return Array.from(npmPkgs);
-}
-
-function buildImportMapFromAllFiles(files: Array<{ file: string; data: string }>) {
-  // Always include react, react-dom/client
-  const importMap: Record<string, string> = {
-    'react': 'https://esm.sh/react@18.2.0',
-    'react-dom/client': 'https://esm.sh/react-dom@18.2.0/client',
-  };
-  const pkgs = getAllNpmImports(files);
-  pkgs.forEach(pkg => {
-    if (pkg === 'react' || pkg === 'react-dom' || pkg === 'react-dom/client') return;
-    importMap[pkg] = `https://esm.sh/${pkg}?deps=react@18.2.0`;
-  });
-  return importMap;
-}
-
-function stripNonNpmImports(code: string, importMap: Record<string, string>) {
-  // Remove all import statements for local files, CSS, and type-only imports
-  // Only keep import statements for npm packages in the import map
-  return code.replace(/import\s+((\w+)|\*\s+as\s+(\w+)|\{([^}]+)\})?\s*from\s*['"]([^'"]+)['"];?|import\s+['"]([^'"]+\.css)['"];?|import\s+type\s+\{[^}]+\}\s+from\s+['"]([^'"]+)['"];?/g, (full, _g1, _d1, _n1, _ni1, pkg, cssImport, typeImport) => {
-    if (cssImport) return '';
-    if (typeImport) return '';
-    if (pkg && (pkg.startsWith('./') || pkg.startsWith('../') || pkg.startsWith('@/'))) return '';
-    if (pkg && importMap[pkg]) return full; // keep npm import
-    return '';
-  });
-}
-
-function filenameToVar(filename: string) {
-  // '@/components/Header' => 'Header', 'app/page.tsx' => 'Home'
-  if (filename === 'app/page.tsx') return 'Home';
-  const parts = filename.split('/');
-  let name = parts[parts.length - 1].replace(/\.[^.]+$/, '');
-  // Capitalize
-  return name.charAt(0).toUpperCase() + name.slice(1);
-}
+import { buildImportMapFromAllFiles } from '@/lib/preview/buildImportMapFromAllFiles';
+import { stripNonNpmImports } from '@/lib/preview/stripNonNpmImports';
+import { filenameToVar } from '@/lib/preview/filenameToVar';
 
 interface StaticPreviewProps {
   onClose: () => void;
